Started adding Health Potions and Health Increases HP Bar unable to display half hearts (aka odd amount of health)

+++ src/com/company/entity/Entity.java
@@ -20,8 +20,8 @@ public class Entity {
     int dialogueIndex= 0;
 
     // CHARACTER STATUS
-    public int maxLife;
-    public int life;
+    public double maxLife;
+    public double life;
 
     public Entity(GamePanel gp) {
         this.gp = gp;
